+++ src/fmt/temporal/printer.rs
@@ -6,13 +6,91 @@ use crate::{
     fmt::{
         buffer::{ArrayBuffer, BorrowedBuffer},
         temporal::{Pieces, PiecesOffset, TimeZoneAnnotationKind},
-        util::{FractionalFormatter, IntegerFormatter},
-        Write, WriteExt,
+        Write,
     },
     span::{Span, UnitSet},
     tz::{Offset, TimeZone},
     SignedDuration, Timestamp, Unit, Zoned,
 };
+
+/// Defines the maximum possible length (in bytes) of an RFC 9557 zoned
+/// datetime.
+///
+/// The actual maximal string possible is I believe
+/// `-009999-03-14T17:30:00.999999999-04:23[America/Argentina/ComodRivadavia]`,
+/// which is 72 bytes. Obviously, the length of the IANA tzdb identifier
+/// matters and can be highly variable.
+///
+/// So why is this bigger than 72? Well, as far as I can tell, there is no
+/// guaranteed maximum length for identifiers. Although, it does appear that
+/// there is some loose goal to keep them as succinct as possible.
+///
+/// So... the current longest identifier is 32 bytes. Surely there will never
+/// be one longer than 50 bytes, right? Well, that's what we assume here.
+/// I don't feel great about this, because if there ever is a longer
+/// identifier, the printer will panic. We can't return an error (here)
+/// because the printer has to be infallible. There are two possible
+/// alternatives here:
+///
+/// 1. We refuse to create any `TimeZone` with an IANA tzdb identifier longer
+/// than the limit we impose here. We can return an error at `TimeZone`
+/// construction (since Jiff already knows how to deal with that) and thus
+/// everyone downstream of `TimeZone`---including this printer---can make
+/// assumptions about the maximum length of identifiers.
+///
+/// 2. Introduce a buffering mechanism that allows us to define a smaller
+/// buffer on the stack that is "flushed" to the caller's `Write` implementation
+/// whenever it fills up. We don't have this abstraction yet (as of 2025-12-31),
+/// but we'll need it for `strftime`.
+///
+/// Perhaps even more pernicious is that this error category will only occur
+/// when writing to a `Write` that doesn't expose an underlying `Vec<u8>`
+/// or `String`. When we have a `String`, and since we can compute a reasonably
+/// tight upper bound on the total length cheaply, we'll always reserve a
+/// sufficient large capacity to write to.
+///
+/// FIXME: So I think we really do need to do either (1) or (2) above. I'm
+/// inclined to do (2) so that we'll handle all possible reasonable inputs.
+/// Although it might be a good idea to impose a maximum limit too. It's very
+/// useful to be able to assume maximum lengths of inputs.
+const MAX_ZONED_LEN: usize = 90;
+
+/// Defines the maximum possible length (in bytes) of an RFC 3339 timestamp
+/// that is always in Zulu time.
+///
+/// The longest possible string is `-009999-03-14T21:53:08.999999999Z`.
+const MAX_TIMESTAMP_ZULU_LEN: usize = 33;
+
+/// Defines the maximum possible length (in bytes) of an RFC 3339 timestamp
+/// with an offset (up to minute precision).
+///
+/// The longest possible string is `-009999-03-15T23:53:07.999999999+25:59`.
+const MAX_TIMESTAMP_OFFSET_LEN: usize = 38;
+
+/// Defines the maximum possible length (in bytes) of an ISO 8601 datetime.
+///
+/// The longest possible string is `-009999-03-14T17:30:00.999999999`.
+const MAX_DATETIME_LEN: usize = 32;
+
+/// Defines the maximum possible length (in bytes) of an ISO 8601 date.
+///
+/// The longest possible string is `-009999-03-14`.
+const MAX_DATE_LEN: usize = 13;
+
+/// Defines the maximum possible length (in bytes) of an ISO 8601 time.
+///
+/// The longest possible string is `17:30:00.999999999`.
+const MAX_TIME_LEN: usize = 18;
+
+/// Defines the maximum possible length (in bytes) of an offset.
+///
+/// The longest possible string is `-25:59:59`.
+const MAX_OFFSET_LEN: usize = 9;
+
+/// Defines the maximum possible length (in bytes) of an ISO 8601 week date.
+///
+/// The longest possible string is `-009999-W11-3`.
+const MAX_ISO_WEEK_DATE_LEN: usize = 13;
 
 /// Defines the maximum possible length (in bytes) of a `Span` printed in the
 /// Temporal ISO 8601 format.
@@ -58,106 +136,196 @@ impl DateTimePrinter {
         DateTimePrinter { precision, ..self }
     }
 
-    pub(super) fn print_zoned<W: Write>(
+    pub(super) fn print_zoned(
         &self,
         zdt: &Zoned,
-        mut wtr: W,
+        wtr: &mut dyn Write,
     ) -> Result<(), Error> {
-        let timestamp = zdt.timestamp();
-        let tz = zdt.time_zone();
-        let offset = tz.to_offset(timestamp);
-        let dt = offset.to_datetime(timestamp);
-        self.print_datetime(&dt, &mut wtr)?;
-        if tz.is_unknown() {
-            wtr.write_str("Z[Etc/Unknown]")?;
-        } else {
-            self.print_offset_rounded(&offset, &mut wtr)?;
-            self.print_time_zone_annotation(&tz, &offset, &mut wtr)?;
+        /// The base size of an RFC 9557 zoned datetime string.
+        ///
+        /// 19 comes from the datetime component, e.g., `2025-01-01T00:00:00`
+        /// 6 comes from the offset, e.g., `+05:30`
+        /// 2 comes from the `[` and `]` for the time zone annotation
+        ///
+        /// Basically, we always need space for *at least* the above stuff.
+        /// The actual time zone annotation name, negative year and fractional
+        /// second component are all optional and can vary quite a bit in
+        /// length.
+        const BASE: usize = 19 + 6 + 2;
+
+        let mut runtime_allocation = BASE
+            + zdt.time_zone().iana_name().map(|name| name.len()).unwrap_or(11);
+        if zdt.year() < 0 {
+            runtime_allocation += 3;
         }
-        Ok(())
+        if zdt.subsec_nanosecond() != 0 {
+            runtime_allocation += 10;
+        }
+
+        BorrowedBuffer::with_writer::<MAX_ZONED_LEN>(
+            wtr,
+            runtime_allocation,
+            |bbuf| Ok(self.print_zoned_impl(zdt, bbuf)),
+        )
     }
 
-    pub(super) fn print_timestamp<W: Write>(
+    fn print_zoned_impl(&self, zdt: &Zoned, bbuf: &mut BorrowedBuffer<'_>) {
+        self.print_datetime_impl(&zdt.datetime(), bbuf);
+        let tz = zdt.time_zone();
+        if tz.is_unknown() {
+            bbuf.write_str("Z[Etc/Unknown]");
+        } else {
+            self.print_offset_rounded(&zdt.offset(), bbuf);
+            self.print_time_zone_annotation(&tz, &zdt.offset(), bbuf);
+        }
+    }
+
+    pub(super) fn print_timestamp(
         &self,
         timestamp: &Timestamp,
-        offset: Option<Offset>,
-        mut wtr: W,
+        wtr: &mut dyn Write,
     ) -> Result<(), Error> {
-        let Some(offset) = offset else {
-            let dt = TimeZone::UTC.to_datetime(*timestamp);
-            self.print_datetime(&dt, &mut wtr)?;
-            self.print_zulu(&mut wtr)?;
-            return Ok(());
-        };
+        let mut runtime_allocation = MAX_TIMESTAMP_ZULU_LEN;
+        if timestamp.subsec_nanosecond() == 0 {
+            runtime_allocation -= 10;
+        }
+        BorrowedBuffer::with_writer::<MAX_TIMESTAMP_ZULU_LEN>(
+            wtr,
+            runtime_allocation,
+            |bbuf| Ok(self.print_timestamp_impl(timestamp, bbuf)),
+        )
+    }
+
+    fn print_timestamp_impl(
+        &self,
+        timestamp: &Timestamp,
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
+        let dt = Offset::UTC.to_datetime(*timestamp);
+        self.print_datetime_impl(&dt, bbuf);
+        self.print_zulu(bbuf);
+    }
+
+    pub(super) fn print_timestamp_with_offset(
+        &self,
+        timestamp: &Timestamp,
+        offset: Offset,
+        wtr: &mut dyn Write,
+    ) -> Result<(), Error> {
+        let mut runtime_allocation = MAX_TIMESTAMP_OFFSET_LEN;
+        if timestamp.subsec_nanosecond() == 0 {
+            runtime_allocation -= 10;
+        }
+        BorrowedBuffer::with_writer::<MAX_TIMESTAMP_OFFSET_LEN>(
+            wtr,
+            runtime_allocation,
+            |bbuf| {
+                Ok(self
+                    .print_timestamp_with_offset_impl(timestamp, offset, bbuf))
+            },
+        )
+    }
+
+    fn print_timestamp_with_offset_impl(
+        &self,
+        timestamp: &Timestamp,
+        offset: Offset,
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
         let dt = offset.to_datetime(*timestamp);
-        self.print_datetime(&dt, &mut wtr)?;
-        self.print_offset_rounded(&offset, &mut wtr)?;
-        Ok(())
+        self.print_datetime_impl(&dt, bbuf);
+        self.print_offset_rounded(&offset, bbuf);
+    }
+
+    pub(super) fn print_datetime(
+        &self,
+        dt: &DateTime,
+        wtr: &mut dyn Write,
+    ) -> Result<(), Error> {
+        let mut runtime_allocation = MAX_DATETIME_LEN;
+        if dt.subsec_nanosecond() == 0 {
+            runtime_allocation -= 10;
+        }
+        BorrowedBuffer::with_writer::<MAX_DATETIME_LEN>(
+            wtr,
+            runtime_allocation,
+            |bbuf| Ok(self.print_datetime_impl(dt, bbuf)),
+        )
     }
 
     /// Formats the given datetime into the writer given.
-    pub(super) fn print_datetime<W: Write>(
+    fn print_datetime_impl(
         &self,
         dt: &DateTime,
-        mut wtr: W,
-    ) -> Result<(), Error> {
-        self.print_date(&dt.date(), &mut wtr)?;
-        wtr.write_char(char::from(if self.lowercase {
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
+        self.print_date_impl(&dt.date(), bbuf);
+        bbuf.write_ascii_char(if self.lowercase {
             self.separator.to_ascii_lowercase()
         } else {
             self.separator
-        }))?;
-        self.print_time(&dt.time(), &mut wtr)?;
-        Ok(())
+        });
+        self.print_time_impl(&dt.time(), bbuf);
+    }
+
+    pub(super) fn print_date(
+        &self,
+        date: &Date,
+        wtr: &mut dyn Write,
+    ) -> Result<(), Error> {
+        BorrowedBuffer::with_writer::<MAX_DATE_LEN>(
+            wtr,
+            MAX_DATE_LEN,
+            |bbuf| Ok(self.print_date_impl(date, bbuf)),
+        )
     }
 
     /// Formats the given date into the writer given.
-    pub(super) fn print_date<W: Write>(
-        &self,
-        date: &Date,
-        mut wtr: W,
-    ) -> Result<(), Error> {
-        static FMT_YEAR_POSITIVE: IntegerFormatter =
-            IntegerFormatter::new().padding(4);
-        static FMT_YEAR_NEGATIVE: IntegerFormatter =
-            IntegerFormatter::new().padding(6);
-        static FMT_TWO: IntegerFormatter = IntegerFormatter::new().padding(2);
-
-        if date.year() >= 0 {
-            wtr.write_int(&FMT_YEAR_POSITIVE, date.year())?;
+    fn print_date_impl(&self, date: &Date, bbuf: &mut BorrowedBuffer<'_>) {
+        let year = date.year();
+        if year >= 0 {
+            bbuf.write_int_pad4(year.unsigned_abs().into());
         } else {
-            wtr.write_int(&FMT_YEAR_NEGATIVE, date.year())?;
+            bbuf.write_ascii_char(b'-');
+            bbuf.write_int_pad6(year.unsigned_abs().into());
         }
-        wtr.write_str("-")?;
-        wtr.write_int(&FMT_TWO, date.month())?;
-        wtr.write_str("-")?;
-        wtr.write_int(&FMT_TWO, date.day())?;
-        Ok(())
+        bbuf.write_ascii_char(b'-');
+        bbuf.write_int_pad2(date.month().unsigned_abs().into());
+        bbuf.write_ascii_char(b'-');
+        bbuf.write_int_pad2(date.day().unsigned_abs().into());
     }
 
     /// Formats the given time into the writer given.
-    pub(super) fn print_time<W: Write>(
+    pub(super) fn print_time(
         &self,
         time: &Time,
-        mut wtr: W,
+        wtr: &mut dyn Write,
     ) -> Result<(), Error> {
-        static FMT_TWO: IntegerFormatter = IntegerFormatter::new().padding(2);
-        static FMT_FRACTION: FractionalFormatter = FractionalFormatter::new();
+        let mut runtime_allocation = MAX_TIME_LEN;
+        if time.subsec_nanosecond() == 0 {
+            runtime_allocation -= 10;
+        }
+        BorrowedBuffer::with_writer::<MAX_TIME_LEN>(
+            wtr,
+            runtime_allocation,
+            |bbuf| Ok(self.print_time_impl(time, bbuf)),
+        )
+    }
 
-        wtr.write_int(&FMT_TWO, time.hour())?;
-        wtr.write_str(":")?;
-        wtr.write_int(&FMT_TWO, time.minute())?;
-        wtr.write_str(":")?;
-        wtr.write_int(&FMT_TWO, time.second())?;
+    fn print_time_impl(&self, time: &Time, bbuf: &mut BorrowedBuffer<'_>) {
+        bbuf.write_int_pad2(time.hour().unsigned_abs().into());
+        bbuf.write_ascii_char(b':');
+        bbuf.write_int_pad2(time.minute().unsigned_abs().into());
+        bbuf.write_ascii_char(b':');
+        bbuf.write_int_pad2(time.second().unsigned_abs().into());
         let fractional_nanosecond = time.subsec_nanosecond();
         if self.precision.map_or(fractional_nanosecond != 0, |p| p > 0) {
-            wtr.write_str(".")?;
-            wtr.write_fraction(
-                &FMT_FRACTION.precision(self.precision),
+            bbuf.write_ascii_char(b'.');
+            bbuf.write_fraction(
+                self.precision,
                 fractional_nanosecond.unsigned_abs(),
-            )?;
+            );
         }
-        Ok(())
     }
 
     /// Formats the given time zone into the writer given.
@@ -166,6 +334,23 @@ impl DateTimePrinter {
         tz: &TimeZone,
         mut wtr: W,
     ) -> Result<(), Error> {
+        // N.B. We use a `&mut dyn Write` here instead of a
+        // `&mut BorrowedBuffer` (as in the other routines for this printer)
+        // because this can emit a POSIX time zone string. We don't really have
+        // strong guarantees about how long this string can be (although all
+        // sensible values are pretty short). Since this API is not expected to
+        // be used much, we don't spend the time to try and optimize this.
+        //
+        // If and when we get a borrowed buffer writer abstraction (for truly
+        // variable length output), then we might consider using that here.
+        self.print_time_zone_impl(tz, &mut wtr)
+    }
+
+    fn print_time_zone_impl(
+        &self,
+        tz: &TimeZone,
+        wtr: &mut dyn Write,
+    ) -> Result<(), Error> {
         if let Some(iana_name) = tz.iana_name() {
             return wtr.write_str(iana_name);
         }
@@ -173,7 +358,12 @@ impl DateTimePrinter {
             return wtr.write_str("Etc/Unknown");
         }
         if let Ok(offset) = tz.to_fixed_offset() {
-            return self.print_offset_full_precision(&offset, wtr);
+            // Kind of unfortunate, but it's probably better than
+            // making `print_offset_full_precision` accept a `dyn Write`.
+            let mut buf = ArrayBuffer::<MAX_OFFSET_LEN>::default();
+            let mut bbuf = buf.as_borrowed();
+            self.print_offset_full_precision(&offset, &mut bbuf);
+            return wtr.write_str(bbuf.filled());
         }
         // We get this on `alloc` because we format the POSIX time zone into a
         // `String` first. See the note below.
@@ -214,88 +404,117 @@ impl DateTimePrinter {
         pieces: &Pieces,
         mut wtr: W,
     ) -> Result<(), Error> {
+        // N.B. We don't bother with writing into the spare capacity of a
+        // `&mut String` here because with `Pieces` it's a little more
+        // complicated to find a more precise upper bound on the length.
+        // Plus, I don't think this API is commonly used, so it's not clear
+        // that it's worth optimizing. But I'm open to a PR with benchmarks
+        // if there's a good use case. ---AG
+        let mut buf = ArrayBuffer::<MAX_ZONED_LEN>::default();
+        let mut bbuf = buf.as_borrowed();
+        // FIXME: This should really use a buffer abstraction. Namely,
+        // the `Pieces` API specifically permits arbitrary length strings
+        // as time zone name annotations. But as written here, this will
+        // panic on overly long strings. Sigh.
+        self.print_pieces_impl(pieces, &mut bbuf);
+        wtr.write_str(bbuf.filled())
+    }
+
+    fn print_pieces_impl(
+        &self,
+        pieces: &Pieces,
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
         if let Some(time) = pieces.time() {
             let dt = DateTime::from_parts(pieces.date(), time);
-            self.print_datetime(&dt, &mut wtr)?;
+            self.print_datetime_impl(&dt, bbuf);
             if let Some(poffset) = pieces.offset() {
-                self.print_pieces_offset(&poffset, &mut wtr)?;
+                self.print_pieces_offset(&poffset, bbuf);
             }
         } else if let Some(poffset) = pieces.offset() {
             // In this case, we have an offset but no time component. Since
             // `2025-01-02-05:00` isn't valid, we forcefully write out the
             // default time (which is what would be assumed anyway).
             let dt = DateTime::from_parts(pieces.date(), Time::midnight());
-            self.print_datetime(&dt, &mut wtr)?;
-            self.print_pieces_offset(&poffset, &mut wtr)?;
+            self.print_datetime_impl(&dt, bbuf);
+            self.print_pieces_offset(&poffset, bbuf);
         } else {
             // We have no time and no offset, so we can just write the date.
             // It's okay to write this followed by an annotation, e.g.,
             // `2025-01-02[America/New_York]` or even `2025-01-02[-05:00]`.
-            self.print_date(&pieces.date(), &mut wtr)?;
+            self.print_date_impl(&pieces.date(), bbuf);
         }
         // For the time zone annotation, a `Pieces` gives us the annotation
         // name or offset directly, where as with `Zoned`, we have a
         // `TimeZone`. So we hand-roll our own formatter directly from the
         // annotation.
         if let Some(ann) = pieces.time_zone_annotation() {
-            wtr.write_str("[")?;
+            bbuf.write_ascii_char(b'[');
             if ann.is_critical() {
-                wtr.write_str("!")?;
+                bbuf.write_ascii_char(b'!');
             }
             match *ann.kind() {
                 TimeZoneAnnotationKind::Named(ref name) => {
-                    wtr.write_str(name.as_str())?
+                    bbuf.write_str(name.as_str());
                 }
                 TimeZoneAnnotationKind::Offset(offset) => {
-                    self.print_offset_rounded(&offset, &mut wtr)?
+                    self.print_offset_rounded(&offset, bbuf);
                 }
             }
-            wtr.write_str("]")?;
+            bbuf.write_ascii_char(b']');
         }
-        Ok(())
     }
 
-    pub(super) fn print_iso_week_date<W: Write>(
+    pub(super) fn print_iso_week_date(
         &self,
         iso_week_date: &ISOWeekDate,
-        mut wtr: W,
+        wtr: &mut dyn Write,
     ) -> Result<(), Error> {
-        static FMT_YEAR_POSITIVE: IntegerFormatter =
-            IntegerFormatter::new().padding(4);
-        static FMT_YEAR_NEGATIVE: IntegerFormatter =
-            IntegerFormatter::new().padding(6);
-        static FMT_TWO: IntegerFormatter = IntegerFormatter::new().padding(2);
-        static FMT_ONE: IntegerFormatter = IntegerFormatter::new().padding(1);
+        BorrowedBuffer::with_writer::<MAX_ISO_WEEK_DATE_LEN>(
+            wtr,
+            MAX_ISO_WEEK_DATE_LEN,
+            |bbuf| Ok(self.print_iso_week_date_impl(iso_week_date, bbuf)),
+        )
+    }
 
-        if iso_week_date.year() >= 0 {
-            wtr.write_int(&FMT_YEAR_POSITIVE, iso_week_date.year())?;
+    fn print_iso_week_date_impl(
+        &self,
+        iso_week_date: &ISOWeekDate,
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
+        let year = iso_week_date.year();
+        if year >= 0 {
+            bbuf.write_int_pad4(year.unsigned_abs().into());
         } else {
-            wtr.write_int(&FMT_YEAR_NEGATIVE, iso_week_date.year())?;
+            bbuf.write_ascii_char(b'-');
+            bbuf.write_int_pad6(year.unsigned_abs().into());
         }
-        wtr.write_str("-")?;
-        wtr.write_char(if self.lowercase { 'w' } else { 'W' })?;
-        wtr.write_int(&FMT_TWO, iso_week_date.week())?;
-        wtr.write_str("-")?;
-        wtr.write_int(
-            &FMT_ONE,
-            iso_week_date.weekday().to_monday_one_offset(),
-        )?;
-        Ok(())
+        bbuf.write_ascii_char(b'-');
+        bbuf.write_ascii_char(if self.lowercase { b'w' } else { b'W' });
+        bbuf.write_int_pad2(iso_week_date.week().unsigned_abs().into());
+        bbuf.write_ascii_char(b'-');
+        bbuf.write_int1(
+            iso_week_date
+                .weekday()
+                .to_monday_one_offset()
+                .unsigned_abs()
+                .into(),
+        );
     }
 
     /// Formats the given "pieces" offset into the writer given.
-    fn print_pieces_offset<W: Write>(
+    fn print_pieces_offset(
         &self,
         poffset: &PiecesOffset,
-        mut wtr: W,
-    ) -> Result<(), Error> {
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
         match *poffset {
-            PiecesOffset::Zulu => self.print_zulu(wtr),
+            PiecesOffset::Zulu => self.print_zulu(bbuf),
             PiecesOffset::Numeric(ref noffset) => {
                 if noffset.offset().is_zero() && noffset.is_negative() {
-                    wtr.write_str("-00:00")
+                    bbuf.write_str("-00:00");
                 } else {
-                    self.print_offset_rounded(&noffset.offset(), wtr)
+                    self.print_offset_rounded(&noffset.offset(), bbuf);
                 }
             }
         }
@@ -305,19 +524,16 @@ impl DateTimePrinter {
     ///
     /// If the given offset has non-zero seconds, then they are rounded to
     /// the nearest minute.
-    fn print_offset_rounded<W: Write>(
+    fn print_offset_rounded(
         &self,
         offset: &Offset,
-        mut wtr: W,
-    ) -> Result<(), Error> {
-        static FMT_TWO: IntegerFormatter = IntegerFormatter::new().padding(2);
-
-        wtr.write_str(if offset.is_negative() { "-" } else { "+" })?;
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
+        bbuf.write_ascii_char(if offset.is_negative() { b'-' } else { b'+' });
         let (offset_hours, offset_minutes) = offset.round_to_nearest_minute();
-        wtr.write_uint(&FMT_TWO, offset_hours)?;
-        wtr.write_str(":")?;
-        wtr.write_uint(&FMT_TWO, offset_minutes)?;
-        Ok(())
+        bbuf.write_int_pad2(offset_hours.into());
+        bbuf.write_ascii_char(b':');
+        bbuf.write_int_pad2(offset_minutes.into());
     }
 
     /// Formats the given offset into the writer given.
@@ -325,33 +541,30 @@ impl DateTimePrinter {
     /// If the given offset has non-zero seconds, then they are emitted as a
     /// third `:`-delimited component of the offset. If seconds are zero, then
     /// only the hours and minute components are emitted.
-    fn print_offset_full_precision<W: Write>(
+    fn print_offset_full_precision(
         &self,
         offset: &Offset,
-        mut wtr: W,
-    ) -> Result<(), Error> {
-        static FMT_TWO: IntegerFormatter = IntegerFormatter::new().padding(2);
-
-        wtr.write_str(if offset.is_negative() { "-" } else { "+" })?;
-        let hours = offset.part_hours_ranged().abs().get();
-        let minutes = offset.part_minutes_ranged().abs().get();
-        let seconds = offset.part_seconds_ranged().abs().get();
-        wtr.write_int(&FMT_TWO, hours)?;
-        wtr.write_str(":")?;
-        wtr.write_int(&FMT_TWO, minutes)?;
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
+        bbuf.write_ascii_char(if offset.is_negative() { b'-' } else { b'+' });
+        let hours = offset.part_hours_ranged().get().unsigned_abs();
+        let minutes = offset.part_minutes_ranged().get().unsigned_abs();
+        let seconds = offset.part_seconds_ranged().get().unsigned_abs();
+        bbuf.write_int_pad2(hours.into());
+        bbuf.write_ascii_char(b':');
+        bbuf.write_int_pad2(minutes.into());
         if seconds > 0 {
-            wtr.write_str(":")?;
-            wtr.write_int(&FMT_TWO, seconds)?;
+            bbuf.write_ascii_char(b':');
+            bbuf.write_int_pad2(seconds.into());
         }
-        Ok(())
     }
 
     /// Prints the "zulu" indicator.
     ///
     /// This should only be used when the offset is not known. For example,
     /// when printing a `Timestamp`.
-    fn print_zulu<W: Write>(&self, mut wtr: W) -> Result<(), Error> {
-        wtr.write_str(if self.lowercase { "z" } else { "Z" })
+    fn print_zulu(&self, bbuf: &mut BorrowedBuffer<'_>) {
+        bbuf.write_ascii_char(if self.lowercase { b'z' } else { b'Z' });
     }
 
     /// Formats the given time zone name into the writer given as an RFC 9557
@@ -361,20 +574,19 @@ impl DateTimePrinter {
     /// is printed instead. (This means the offset will be printed twice, which
     /// is indeed an intended behavior of RFC 9557 for cases where a time zone
     /// name is not used or unavailable.)
-    fn print_time_zone_annotation<W: Write>(
+    fn print_time_zone_annotation(
         &self,
         time_zone: &TimeZone,
         offset: &Offset,
-        mut wtr: W,
-    ) -> Result<(), Error> {
-        wtr.write_str("[")?;
+        bbuf: &mut BorrowedBuffer<'_>,
+    ) {
+        bbuf.write_ascii_char(b'[');
         if let Some(iana_name) = time_zone.iana_name() {
-            wtr.write_str(iana_name)?;
+            bbuf.write_str(iana_name);
         } else {
-            self.print_offset_rounded(offset, &mut wtr)?;
+            self.print_offset_rounded(offset, bbuf);
         }
-        wtr.write_str("]")?;
-        Ok(())
+        bbuf.write_ascii_char(b']');
     }
 }
 
@@ -685,7 +897,7 @@ mod tests {
         let zoned: Zoned = dt.in_tz("America/New_York").unwrap();
         let mut buf = String::new();
         DateTimePrinter::new()
-            .print_timestamp(&zoned.timestamp(), None, &mut buf)
+            .print_timestamp(&zoned.timestamp(), &mut buf)
             .unwrap();
         assert_eq!(buf, "2024-03-10T09:34:45Z");
 
@@ -693,9 +905,40 @@ mod tests {
         let zoned: Zoned = dt.in_tz("America/New_York").unwrap();
         let mut buf = String::new();
         DateTimePrinter::new()
-            .print_timestamp(&zoned.timestamp(), None, &mut buf)
+            .print_timestamp(&zoned.timestamp(), &mut buf)
             .unwrap();
         assert_eq!(buf, "-002024-03-10T10:30:47Z");
+    }
+
+    #[test]
+    fn print_timestamp_with_offset() {
+        if crate::tz::db().is_definitively_empty() {
+            return;
+        }
+
+        let dt = date(2024, 3, 10).at(5, 34, 45, 0);
+        let zoned: Zoned = dt.in_tz("America/New_York").unwrap();
+        let mut buf = String::new();
+        DateTimePrinter::new()
+            .print_timestamp_with_offset(
+                &zoned.timestamp(),
+                zoned.offset(),
+                &mut buf,
+            )
+            .unwrap();
+        assert_eq!(buf, "2024-03-10T05:34:45-04:00");
+
+        let dt = date(-2024, 3, 10).at(5, 34, 45, 0);
+        let zoned: Zoned = dt.in_tz("America/New_York").unwrap();
+        let mut buf = String::new();
+        DateTimePrinter::new()
+            .print_timestamp_with_offset(
+                &zoned.timestamp(),
+                zoned.offset(),
+                &mut buf,
+            )
+            .unwrap();
+        assert_eq!(buf, "-002024-03-10T05:34:45-04:56");
     }
 
     #[test]